backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMuPDF==1.23.26
python-dotenv==1.0.0
openai==1.12.0
Werkzeug==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.7

uuid==1.30